Restructuring integration tests
DCOPS-7239
