+++ sous/test/actual_deploy_test.go
@@ -3,25 +3,24 @@ package test
 import (
 	"flag"
 	"fmt"
-	"log"
-	"net"
 	"os"
-	"regexp"
 	"testing"
-	"time"
 
-	"github.com/opentable/singularity"
-	"github.com/opentable/singularity/dtos"
 	"github.com/opentable/sous/sous"
-	"github.com/opentable/sous/test_with_docker"
 	"github.com/opentable/sous/util/docker_registry"
-	"github.com/satori/go.uuid"
+	"github.com/samsalisbury/semv"
 	"github.com/stretchr/testify/assert"
 )
 
-var ip, registryName, imageName, singularityURL string
+var imageName string
+
+func TestMain(m *testing.M) {
+	flag.Parse()
+	os.Exit(wrapCompose(m))
+}
 
 func TestGetLabels(t *testing.T) {
+	registerLabelledContainers()
 	assert := assert.New(t)
 	cl := docker_registry.NewClient()
 	cl.BecomeFoolishlyTrusting()
@@ -30,10 +29,13 @@ func TestGetLabels(t *testing.T) {
 
 	assert.Nil(err)
 	assert.Contains(labels, sous.DockerRepoLabel)
+	resetSingularity()
 }
 
 func TestGetRunningDeploymentSet(t *testing.T) {
 	assert := assert.New(t)
+
+	registerLabelledContainers()
 
 	deps, err := sous.GetRunningDeploymentSet([]string{singularityURL})
 	assert.Nil(err)
@@ -55,133 +57,70 @@ func TestGetRunningDeploymentSet(t *testing.T) {
 	assert.Equal("91495f1b1630084e301241100ecf2e775f6b672c", grafana.SourceVersion.Meta)
 	assert.Equal(1, grafana.NumInstances)
 	assert.Equal(sous.ManifestKindService, grafana.Kind)
+
+	resetSingularity()
 }
 
-func TestMain(m *testing.M) {
-	flag.Parse()
-	os.Exit(wrapCompose(m))
+func TestResolve(t *testing.T) {
+	assert := assert.New(t)
+
+	clusterDefs := sous.Defs{
+		Clusters: sous.Clusters{
+			"it": sous.Cluster{
+				BaseURL: singularityURL,
+			},
+		},
+	}
+
+	stateOneTwo := sous.State{
+		Defs: clusterDefs,
+		Manifests: sous.Manifests{
+			"one": manifest("https://github.com/opentable/one", "1.1.1"),
+			"two": manifest("https://github.com/opentable/two", "1.1.1"),
+		},
+	}
+	stateTwoThree := sous.State{
+		Defs: clusterDefs,
+		Manifests: sous.Manifests{
+			"two":   manifest("https://github.com/opentable/two", "1.1.1"),
+			"three": manifest("https://github.com/opentable/three", "1.1.1"),
+		},
+	}
+
+	Resolve(stateOneTwo)
+	// one and two are running
+	Resolve(stateTwoThree)
+	// two and three are running, not one
+
+	resetSingularity()
 }
 
-func wrapCompose(m *testing.M) (resultCode int) {
-	log.SetFlags(log.Flags() | log.Lshortfile)
-
-	if testing.Short() {
-		return 0
+func manifest(sourceURL, version string) sous.Manifest {
+	return sous.Manifest{
+		Source: sous.SourceLocation{
+			RepoURL:    sous.RepoURL(sourceURL),
+			RepoOffset: sous.RepoOffset(""),
+		},
+		Owners: []string{`xyz`},
+		Kind:   sous.ManifestKindService,
+		Deployments: sous.DeploySpecs{
+			"it": sous.PartialDeploySpec{
+				DeployConfig: sous.DeployConfig{
+					Resources:    sous.Resources{}, //map[string]string
+					Args:         []string{},
+					Env:          sous.Env{}, //map[s]s
+					NumInstances: 1,
+				},
+				Version: semv.MustParse(version),
+				//clusterName: "it",
+			},
+		},
 	}
+}
 
-	defer func() {
-		log.Println("Cleaning up...")
-		if err := recover(); err != nil {
-			log.Print("Panic: ", err)
-			resultCode = 1
-		}
-	}()
-
-	testAgent, err := test_with_docker.NewAgentWithTimeout(5 * time.Minute)
-	if err != nil {
-		panic(err)
-	}
-
-	ip, err := testAgent.IP()
-	if err != nil {
-		panic(err)
-	}
-
-	composeDir := "test-registry"
-	registryCertName := "testing.crt"
-	registryName = fmt.Sprintf("%s:%d", ip, 5000)
-
-	err = registryCerts(testAgent, composeDir, registryCertName, ip)
-	if err != nil {
-		panic(fmt.Errorf("building registry certs failed: %s", err))
-	}
-
-	started, err := testAgent.ComposeServices(composeDir, map[string]uint{"Singularity": 7099, "Registry": 5000})
-	defer testAgent.Shutdown(started)
-
+func registerLabelledContainers() {
 	registerAndDeploy(ip, "hello-labels", "hello-labels", []int32{})
 	registerAndDeploy(ip, "hello-server-labels", "hello-server-labels", []int32{80})
 	registerAndDeploy(ip, "grafana-repo", "grafana-labels", []int32{})
 	imageName = fmt.Sprintf("%s/%s:%s", registryName, "grafana-repo", "latest")
-
-	log.Print("   *** Beginning tests... ***\n\n")
-	resultCode = m.Run()
-	return
-}
-
-func registerAndDeploy(ip net.IP, reponame, dir string, ports []int32) (err error) {
-	imageName := fmt.Sprintf("%s/%s:%s", registryName, reponame, "latest")
-	err = buildAndPushContainer(dir, imageName)
-	if err != nil {
-		panic(fmt.Errorf("building test container failed: %s", err))
-	}
-
-	singularityURL = fmt.Sprintf("http://%s:%d/singularity", ip, 7099)
-	err = startInstance(singularityURL, imageName, ports)
-	if err != nil {
-		panic(fmt.Errorf("starting a singularity instance failed: %s", err))
-	}
-
-	return
-}
-
-type dtoMap map[string]interface{}
-
-func loadMap(fielder dtos.Fielder, m dtoMap) dtos.Fielder {
-	_, err := dtos.LoadMap(fielder, m)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	return fielder
-}
-
-var notInIDre = regexp.MustCompile(`[-/]`)
-
-func idify(in string) string {
-	return notInIDre.ReplaceAllString(in, "")
-}
-
-func startInstance(url, imageName string, ports []int32) error {
-	reqID := idify(imageName)
-
-	sing := singularity.NewClient(url)
-
-	req := loadMap(&dtos.SingularityRequest{}, map[string]interface{}{
-		"Id":          reqID,
-		"RequestType": dtos.SingularityRequestRequestTypeSERVICE,
-		"Instances":   int32(1),
-	}).(*dtos.SingularityRequest)
-
-	_, err := sing.PostRequest(req)
-	if err != nil {
-		return err
-	}
-
-	dockerInfo := loadMap(&dtos.SingularityDockerInfo{}, dtoMap{
-		"Image": imageName,
-	}).(*dtos.SingularityDockerInfo)
-
-	depReq := loadMap(&dtos.SingularityDeployRequest{}, dtoMap{
-		"Deploy": loadMap(&dtos.SingularityDeploy{}, dtoMap{
-			"Id":        idify(uuid.NewV4().String()),
-			"RequestId": reqID,
-			"Resources": loadMap(&dtos.Resources{}, dtoMap{
-				"Cpus":     0.1,
-				"MemoryMb": 100.0,
-				"NumPorts": int32(1),
-			}),
-			"ContainerInfo": loadMap(&dtos.SingularityContainerInfo{}, dtoMap{
-				"Type":   dtos.SingularityContainerInfoSingularityContainerTypeDOCKER,
-				"Docker": dockerInfo,
-			}),
-		}),
-	}).(*dtos.SingularityDeployRequest)
-
-	_, err = sing.Deploy(depReq)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
